backend/requirements.txt
torch==1.10.0
torchvision==0.11.0
pillow==8.4.0
flask==2.0.2
flask-cors==3.0.10
requests==2.27.1